refactor(schemas): migrate Phase 28 endpoint schemas — 2 files, 8 classes + 2 enums (#6042)
- secrets.py: migrate SecretScope, SecretType enums + SecretModel, SecretCreateRequest
  (with helper methods to_secret_model/is_chat_scoped/etc.), SecretUpdateRequest,
  SecretTransferRequest → schemas_system.py
- enhanced_search.py: migrate SearchRequest→NPUSearchRequest, SearchResponse→
  NPUSearchResponse, BenchmarkRequest, OptimizationRequest→NPUOptimizationRequest
  → schemas_knowledge.py

Renames to avoid generic-name collisions:
- SearchRequest → NPUSearchRequest (knowledge has SearchRequest)
- SearchResponse → NPUSearchResponse (knowledge has SearchResponse)
- OptimizationRequest → NPUOptimizationRequest (agent has LLMOptimizationRequest)

schemas_system.py: gains uuid + _SECRET_NAME_PATTERN regex + _validate_secret_name
helper for SecretCreate/Update validators.
