+++ autobot-backend/api/secrets.py
@@ -20,19 +20,16 @@ import logging
 import os
 import re
 import threading
-import uuid
 from collections import defaultdict
 from copy import deepcopy
 from datetime import datetime, timezone
 from autobot_shared.time_utils import parse_utc_iso
-from enum import Enum
 from time import time
 from typing import Dict, List, Optional
 
 from cryptography.fernet import Fernet
 from fastapi import APIRouter, Depends, HTTPException, Query, Request
 from fastapi.responses import JSONResponse
-from pydantic import BaseModel, Field, field_validator
 
 from auth_middleware import check_admin_permission, get_auth_middleware
 from autobot_memory_graph import AutoBotMemoryGraph
@@ -41,7 +38,15 @@ from autobot_shared.error_boundaries import ErrorCategory, with_error_handling
 from middleware.proxy_utils import get_client_ip
 from type_defs.common import Metadata
 from api.schemas_common import DataResponse
-from api.schemas_system import SecretsStatusResponse
+from api.schemas_system import (
+    SecretCreateRequest,
+    SecretModel,
+    SecretScope,
+    SecretTransferRequest,
+    SecretType,
+    SecretUpdateRequest,
+    SecretsStatusResponse,
+)
 
 logger = logging.getLogger(__name__)
 
@@ -101,139 +106,6 @@ def validate_secret_name(name: str) -> str:
             "underscores, hyphens, and dots"
         )
     return name
-
-
-class SecretScope(str, Enum):
-    """Secret scope enumeration (Issue #685: aligned with database model)"""
-
-    CHAT = "chat"
-    GENERAL = "general"
-    USER = "user"  # Private to user
-    SESSION = "session"  # Session-scoped
-    SHARED = "shared"  # Explicitly shared
-    GROUP = "group"  # Team members
-    ORGANIZATION = "organization"  # All org members
-
-
-class SecretType(str, Enum):
-    """Secret type enumeration"""
-
-    SSH_KEY = "ssh_key"
-    PASSWORD = "password"  # nosec B105 - enum value, not actual password
-    API_KEY = "api_key"
-    TOKEN = "token"  # nosec B105 - enum value, not actual token
-    CERTIFICATE = "certificate"
-    DATABASE_URL = "database_url"
-    INFRASTRUCTURE_HOST = "infrastructure_host"  # SSH/VNC host credentials
-    OTHER = "other"
-
-
-class SecretModel(BaseModel):
-    """Secret data model"""
-
-    id: str = Field(default_factory=lambda: str(uuid.uuid4()))
-    name: str
-    type: SecretType
-    scope: SecretScope
-    chat_id: Optional[str] = None
-    description: Optional[str] = ""
-    tags: List[str] = Field(default_factory=list)
-    created_at: datetime = Field(default_factory=datetime.now)
-    updated_at: datetime = Field(default_factory=datetime.now)
-    expires_at: Optional[datetime] = None
-    metadata: Metadata = Field(default_factory=dict)
-
-
-class SecretCreateRequest(BaseModel):
-    """Request model for creating secrets (Issue #685: hierarchical access)"""
-
-    name: str = Field(..., min_length=1, max_length=256)
-    type: SecretType
-    scope: SecretScope
-    value: str = Field(..., min_length=1, max_length=65536)  # 64KB max
-    chat_id: Optional[str] = Field(None, max_length=128)
-    description: Optional[str] = Field("", max_length=1024)
-    tags: List[str] = Field(default_factory=list)
-    expires_at: Optional[datetime] = None
-    metadata: Metadata = Field(default_factory=dict)
-    # Issue #685: Hierarchical access fields
-    owner_id: Optional[str] = Field(None, max_length=128, description="Owner user ID")
-    org_id: Optional[str] = Field(
-        None, max_length=128, description="Organization ID for org-level secrets"
-    )
-    team_ids: List[str] = Field(
-        default_factory=list, description="Team IDs for group-level secrets"
-    )
-    shared_with: List[str] = Field(
-        default_factory=list, description="User IDs to share with"
-    )
-
-    @field_validator("name")
-    @classmethod
-    def validate_name(cls, v: str) -> str:
-        """Validate secret name contains only safe characters"""
-        return validate_secret_name(v)
-
-    # === Issue #372: Feature Envy Reduction Methods ===
-
-    def to_secret_model(self, secret_id: Optional[str] = None) -> "SecretModel":
-        """Convert request to SecretModel (Issue #372 - reduces feature envy).
-
-        Args:
-            secret_id: Optional ID for the secret, generates UUID if not provided
-
-        Returns:
-            SecretModel instance with request data
-        """
-        return SecretModel(
-            id=secret_id or str(uuid.uuid4()),
-            name=self.name,
-            type=self.type,
-            scope=self.scope,
-            chat_id=self.chat_id if self.scope == SecretScope.CHAT else None,
-            description=self.description,
-            tags=self.tags,
-            expires_at=self.expires_at,
-            metadata=self.metadata,
-        )
-
-    def is_chat_scoped(self) -> bool:
-        """Check if secret is chat-scoped (Issue #372 - reduces feature envy)."""
-        return self.scope == SecretScope.CHAT
-
-    def requires_chat_id(self) -> bool:
-        """Check if chat_id is required but missing (Issue #372)."""
-        return self.is_chat_scoped() and not self.chat_id
-
-    def get_log_summary(self) -> str:
-        """Get formatted log summary (Issue #372 - reduces feature envy)."""
-        return f"{self.scope.value} secret '{self.name}'"
-
-
-class SecretUpdateRequest(BaseModel):
-    """Request model for updating secrets"""
-
-    name: Optional[str] = Field(None, min_length=1, max_length=256)
-    description: Optional[str] = Field(None, max_length=1024)
-    tags: Optional[List[str]] = None
-    expires_at: Optional[datetime] = None
-    metadata: Optional[Metadata] = None
-
-    @field_validator("name")
-    @classmethod
-    def validate_name(cls, v: Optional[str]) -> Optional[str]:
-        """Validate secret name contains only safe characters"""
-        if v is not None:
-            return validate_secret_name(v)
-        return v
-
-
-class SecretTransferRequest(BaseModel):
-    """Request model for transferring secrets between scopes"""
-
-    secret_ids: List[str]
-    target_scope: SecretScope
-    target_chat_id: Optional[str] = None
 
 
 class SecretsManager:
